Allow telemetry 1.0 to be used
Telemetry 1.0.0 is pretty much the same as 0.4.3 and was made to mark
the telemetry API as stable. This allows both 0.4 and 1.0 releases to be
considered when hex evaluates project dependencies. This avoids
downgrades of telemetry_poller and cowboy_telemetry and possibly other
libraries.

See https://diff.hex.pm/diff/telemetry/0.4.3..1.0.0 for the changes
between 0.4.3 and 1.0.0.