fix examples for c:Slipstream.handle_disconnect/2
Slipstream.reconnect/1 returns an ok or error tuple, so wrapping
reconnect in its own {:ok, socket} tuple (as in line 534) is a
double wrapping. This is also a good opportunity to document stopping
the GenServer on failure to reconnect rather than giving a MatchError.

Thanks to @cadebward for reporting this!

closes #47