telemetry: Trim metadata from sockets when constructing metadata
The Slipstream.Socket struct contained the metadata used for telemetry
events like opening/closing connections and joining/leaving. The
metadata also contained a prior version of the socket though, and each attempt to join the topic would extend this tree by another level.
So repeated failures to connect or join a topic could cause the socket
to grow rapidly based on the number of retries and the number of
topics being (re)joined.

To fix this, we trim the socket's metadata field when storing a socket
in metadata. The metadata storage in the socket is redundant in the
telemetry message because the rest of the telemetry message contains
the metadata.