Add support for a binary serializer
In some cases, when using Phoenix channels, it is possible to use a
binary serializer. The library already has support for binary messages,
but this is slightly different. With a binary serializer, all messages
can be sent in a binary format, not just individual payloads. This
commit changes the contract for the serializer to allow returning a
binary as before, or the tuple `{:binary, binary}`

In the integration tests, an example binary encoder that uses
`term_to_binary/1` and `binary_to_term/1` is used. The integration tests
are run both against the existing text based serializer, and the
reference binary serializer. The `GoodExample` has been modified to take
a set of options instead of reading them from the config, although a
fallback is used to minimize the number of tests updated.