Fix broken `decode_binary!`
Encoding was correct, but the pattern match for decoding wasn't
properly accounting for a couple of values that Phoenix uses.

Added a round trip test for binary encoding to verify that encode/decode
returns the original value. Also added a round trip test for non-binary
json encoding for good measure (even though it wasn't broken).